backend/requirements.txt
alembic==1.13.1
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.2
flask-marshmallow==1.2.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
frozendict==2.3.8
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Mako==1.3.2
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
multitasking==0.0.11
numpy==1.26.1
packaging==23.2
pandas==2.1.3
peewee==3.17.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1
Werkzeug==2.3.7
yfinance==0.2.35